// ClientApp/src/types/Meta.ts
export class Meta {
  title?: string;
  alternateTitles?: string;
  library?: string;
  developer?: string;
  publisher?: string;
  series?: string;
  source?: string;
  tags?: string;
  tagCategories?: string;
  playMode?: string;
  releaseDate?: string;
  version?: string;
  languages?: string;
  extreme?: string;
  status?: string;
  notes?: string;
  originalDescription?: string;
  curationNotes?: string;
  applicationPath?: string;
  launchCommand?: string;
}